requirements.txt
PyYAML==6.0
astropy==5.1
git+https://gitlab.com/free-astro/pysiril.git@V0_0_14
